requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.2
django-import-export==4.3.7
django-restframework==0.0.1
djangorestframework==3.16.0
pillow==11.2.1
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
whitenoise==6.9.0